# TourPlanner.API/TourPlanner.API.swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "TourPlanner.API",
    "version": "1.0"
  },
  "paths": {
    "/Tour": {
      "get": {
        "tags": [
          "Tour"
        ],
        "operationId": "GetTours",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TourDomain"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TourDomain"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TourDomain"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tour"
        ],
        "operationId": "AddNewTour",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TourDomain"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TourDomain"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TourDomain"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TourDomain"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourDomain"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourDomain"
                }
              }
            }
          }
        }
      }
    },
    "/Tour/{id}": {
      "delete": {
        "tags": [
          "Tour"
        ],
        "operationId": "DeleteTour",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Tour"
        ],
        "operationId": "UpdateTour",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TourDomain"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TourDomain"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TourDomain"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TourDomain"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourDomain"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourDomain"
                }
              }
            }
          }
        }
      }
    },
    "/TourLog/{id}": {
      "get": {
        "tags": [
          "TourLog"
        ],
        "operationId": "GetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TourLogDomain"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourLogDomain"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourLogDomain"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "TourLog"
        ],
        "operationId": "DeleteTourLog",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "TourLog"
        ],
        "operationId": "UpdateTourLog",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TourLogDomain"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TourLogDomain"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TourLogDomain"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TourLogDomain"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourLogDomain"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourLogDomain"
                }
              }
            }
          }
        }
      }
    },
    "/TourLog": {
      "post": {
        "tags": [
          "TourLog"
        ],
        "operationId": "AddNewTourLog",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TourLogDomain"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TourLogDomain"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TourLogDomain"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TourLogDomain"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourLogDomain"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourLogDomain"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "TourDomain": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "from": {
            "type": "string",
            "nullable": true
          },
          "to": {
            "type": "string",
            "nullable": true
          },
          "geoJson": {
            "type": "string",
            "nullable": true
          },
          "directions": {
            "type": "string",
            "nullable": true
          },
          "distance": {
            "type": "number",
            "format": "double"
          },
          "estimatedTime": {
            "type": "number",
            "format": "double"
          },
          "transportType": {
            "type": "string",
            "nullable": true
          },
          "tourLogs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TourLogDomain"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TourLogDomain": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "comment": {
            "type": "string",
            "nullable": true
          },
          "totalDistance": {
            "type": "number",
            "format": "double"
          },
          "totalTime": {
            "type": "number",
            "format": "double"
          },
          "difficulty": {
            "type": "number",
            "format": "double"
          },
          "rating": {
            "type": "number",
            "format": "double"
          },
          "tourId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      }
    }
  }
}